+++ py-polars/tests/unit/utils/test_utils.py
@@ -1,7 +1,5 @@
 from __future__ import annotations
 
-import inspect
-import warnings
 from datetime import date, datetime, time, timedelta
 from typing import TYPE_CHECKING, Any
 
@@ -15,7 +13,6 @@ from polars.utils.convert import (
     _timedelta_to_pl_duration,
     _timedelta_to_pl_timedelta,
 )
-from polars.utils.deprecation import deprecate_nonkeyword_arguments, redirect
 from polars.utils.meta import get_idx_type
 from polars.utils.various import _in_notebook, parse_version
 
@@ -117,50 +114,6 @@ def test_estimated_size() -> None:
 def test_parse_version(v1: Any, v2: Any) -> None:
     assert parse_version(v1) > parse_version(v2)
     assert parse_version(v2) < parse_version(v1)
-
-
-class Foo:  # noqa: D101
-    @deprecate_nonkeyword_arguments(allowed_args=["self", "baz"])
-    def bar(  # noqa: D102
-        self, baz: str, ham: str | None = None, foobar: str | None = None
-    ) -> None:
-        ...
-
-
-def test_deprecate_nonkeyword_arguments_method_signature() -> None:
-    # Note the added star indicating keyword-only arguments after 'baz'
-    expected = "(self, baz: 'str', *, ham: 'str | None' = None, foobar: 'str | None' = None) -> 'None'"
-    assert str(inspect.signature(Foo.bar)) == expected
-
-
-def test_deprecate_nonkeyword_arguments_method_warning() -> None:
-    msg = (
-        r"All arguments of Foo\.bar except for \'baz\' will be keyword-only in the next breaking release."
-        r" Use keyword arguments to silence this warning."
-    )
-    with pytest.deprecated_call(match=msg):
-        Foo().bar("qux", "quox")
-
-
-def test_redirect() -> None:
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore", DeprecationWarning)
-
-        # one-to-one redirection
-        @redirect({"foo": "bar"})
-        class DemoClass1:
-            def bar(self, upper: bool = False) -> str:
-                return "BAZ" if upper else "baz"
-
-        assert DemoClass1().foo() == "baz"  # type: ignore[attr-defined]
-
-        # redirection with **kwargs
-        @redirect({"foo": ("bar", {"upper": True})})
-        class DemoClass2:
-            def bar(self, upper: bool = False) -> str:
-                return "BAZ" if upper else "baz"
-
-        assert DemoClass2().foo() == "BAZ"  # type: ignore[attr-defined]
 
 
 def test_get_idx_type_deprecation() -> None:
